requirements.txt
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
pytest==7.4.3
